+++ flake.nix
@@ -35,7 +35,7 @@
           nixpkgs.overlays = import ./overlays;
         };
 
-      osConfigAarch64 =
+      rockchipOsConfigAarch64 =
         buildPlatform: deviceModule: variant:
         nixpkgs.lib.nixosSystem {
           system = "aarch64-linux";
@@ -52,7 +52,7 @@
           ];
         };
 
-      installerConfigAarch64 =
+      rockchipInstallerConfigAarch64 =
         buildPlatform: deviceModule:
         nixpkgs.lib.nixosSystem {
           system = "aarch64-linux";
@@ -62,6 +62,37 @@
             { nixpkgs.hostPlatform = "aarch64-linux"; }
             { nixpkgs.buildPlatform = buildPlatform; }
             rockchip.nixosModules.sdImageRockchipInstaller
+            home-manager.nixosModules.home-manager
+            ./config.nix
+            deviceModule
+          ];
+        };
+
+      osConfigAarch64 =
+        buildPlatform: deviceModule: variant:
+        nixpkgs.lib.nixosSystem {
+          system = "aarch64-linux";
+          specialArgs = { inherit buildPlatform settings; };
+          modules = [
+            overlayModule
+            { nixpkgs.hostPlatform = "aarch64-linux"; }
+            { nixpkgs.buildPlatform = buildPlatform; }
+            home-manager.nixosModules.home-manager
+            ./config.nix
+            variant
+            deviceModule
+          ];
+        };
+
+      installerConfigAarch64 =
+        buildPlatform: deviceModule:
+        nixpkgs.lib.nixosSystem {
+          system = "aarch64-linux";
+          specialArgs = { inherit buildPlatform settings; };
+          modules = [
+            overlayModule
+            { nixpkgs.hostPlatform = "aarch64-linux"; }
+            { nixpkgs.buildPlatform = buildPlatform; }
             home-manager.nixosModules.home-manager
             ./config.nix
             deviceModule
@@ -97,21 +128,27 @@
       };
 
       nixosConfigurations.PineBookPro =
-        osConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
           ./gnome.nix;
       nixosConfigurations.PineBookPro-plasma =
-        osConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
           ./plasma.nix;
       nixosConfigurations.PineBookPro-phosh =
-        osConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinebook-pro.nix
           ./phosh.nix;
-      nixosConfigurations.PineTab2 = osConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix ./gnome.nix;
+      nixosConfigurations.PineTab2 =
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix
+          ./gnome.nix;
       nixosConfigurations.PineTab2-plasma =
-        osConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix
           ./plasma.nix;
       nixosConfigurations.PineTab2-phosh =
-        osConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix
+        rockchipOsConfigAarch64 "x86_64-linux" ./devices/pinetab2.nix
           ./phosh.nix;
+      nixosConfigurations.GenericAarch64 =
+        osConfigAarch64 "x86_64-linux" ./devices/generic-aarch64.nix
+          ./gnome.nix;
+      nixosConfigurations.GenericAarch64-installer = installerConfigAarch64 "x86_64-linux" ./devices/generic-aarch64.nix;
     }
     // utils.lib.eachDefaultSystem (
       system:
@@ -123,6 +160,7 @@
             lib
             pkgs
             system
+            rockchipOsConfigAarch64
             osConfigAarch64
             installerConfigX86
             ;
@@ -132,6 +170,8 @@
             lib
             pkgs
             system
+            rockchipOsConfigAarch64
+            rockchipInstallerConfigAarch64
             osConfigAarch64
             installerConfigAarch64
             installerConfigX86